requirement.txt
torch==2.0.0
torchvision==0.15.1
transformers==4.30.0
easyocr==1.4
opencv-python==4.7.0.72
scikit-learn==1.2.2
tqdm==4.65.0
Pillow==9.5.0
